requirements.txt
altgraph                  0.17.2
click                     8.0.1
colorama                  0.4.4
future                    0.18.2
importlib-metadata        4.8.1
numpy                     1.21.3
pefile                    2021.9.3
Pillow                    8.3.2
pip                       10.0.1
pyinstaller               4.5.1
pyinstaller-hooks-contrib 2021.3
PyQt5                     5.15.4
pyqt5-plugins             5.15.4.2.2
PyQt5-Qt5                 5.15.2
PyQt5-sip                 12.9.0
pyqt5-tools               5.15.4.3.2
pyqtgraph                 0.12.3
python-dotenv             0.19.0
pywin32-ctypes            0.2.0
qt5-applications          5.15.2.2.2
qt5-tools                 5.15.2.1.2
setuptools                39.0.1
typing-extensions         3.10.0.2
ui                        0.1.4
zipp                      3.5.0
